Add endpoint to delete videos
Make it a little easier for local development and testing. Since
we have de-duplication in place based on md5 hash of uploaded files,
it's a bit of a pain to test locally. Add a delete endpoint to be able
to delete a previously uploaded file so that we can upload the same
file again.